refactor: replace bandwidth_monitor's format_bytes with a new utility function for better code organization
